+++ quiz1/list.h
@@ -5,13 +5,13 @@
 typedef struct __node {
     struct __node *left, *right;
     struct __node *next;
-    int *value;
+    uint32_t *value;
 } node_t;
 
 void __list_add(node_t **, node_t *);
 node_t *__list_tail(node_t **);
-int __list_length(node_t **);
-node_t *__list_construct(node_t *, int *);
+size_t __list_length(node_t **);
+node_t *__list_construct(node_t *, uint32_t *);
 void __list_free(node_t **);
 void __print_list(node_t *);
 void __quick_sort(node_t **);
